app/requirements.txt
fastapi
uvicorn
pandas
sqlalchemy
psycopg2-binary
requests
python-dotenv
cryptography
numpy
